Not a hard problem but found a crazy solution that does increment in O(1) time.
The idea is to do increment lazily. We have an array where inc[i] stores the increment
for all elements with index <= i. When we pop, we drag along this i accordingly.
Idea was taken from the leetcode god @lee215.

wow github copilot can auto complete lee215 for me, that op huh